use optparse to parse default command in toml config; needed to add Parse.hs-boot to resolve cyclic dependency
