+++ docs/doc-index-P.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd"><html xmlns="http://www.w3.org/1999/xhtml"><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8" /><meta name="viewport" content="width=device-width, initial-scale=1" /><title>hoyo-0.3.0.0: Bookmark directories for cd (Index - P)</title><link href="linuwial.css" rel="stylesheet" type="text/css" title="Linuwial" /><link rel="stylesheet" type="text/css" href="quick-jump.css" /><link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=PT+Sans:400,400i,700" /><script src="haddock-bundle.min.js" async="async" type="text/javascript"></script><script type="text/x-mathjax-config">MathJax.Hub.Config({ tex2jax: { processClass: "mathjax", ignoreClass: ".*" } });</script><script src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-AMS-MML_HTMLorMML" type="text/javascript"></script></head><body><div id="package-header"><span class="caption">hoyo-0.3.0.0: Bookmark directories for cd</span><ul class="links" id="page-menu"><li><a href="index.html">Contents</a></li><li><a href="doc-index.html">Index</a></li></ul></div><div id="content"><div id="alphabet"><ul><li><a href="doc-index-A.html">A</a></li><li><a href="doc-index-B.html">B</a></li><li><a href="doc-index-C.html">C</a></li><li><a href="doc-index-D.html">D</a></li><li><a href="doc-index-E.html">E</a></li><li><a href="doc-index-F.html">F</a></li><li><a href="doc-index-G.html">G</a></li><li><a href="doc-index-H.html">H</a></li><li><a href="doc-index-I.html">I</a></li><li><a href="doc-index-L.html">L</a></li><li><a href="doc-index-M.html">M</a></li><li><a href="doc-index-N.html">N</a></li><li><a href="doc-index-O.html">O</a></li><li><a href="doc-index-P.html">P</a></li><li><a href="doc-index-R.html">R</a></li><li><a href="doc-index-S.html">S</a></li><li><a href="doc-index-T.html">T</a></li><li><a href="doc-index-U.html">U</a></li><li><a href="doc-index-V.html">V</a></li><li><a href="doc-index-W.html">W</a></li><li><a href="doc-index-95.html">_</a></li><li><a href="doc-index-All.html">All</a></li></ul></div><div id="index"><p class="caption">Index - P</p><table><tr><td class="src">parseCommand</td><td class="module"><a href="HoYo-CLI-Parse.html#v:parseCommand" title="HoYo.CLI.Parse">HoYo.CLI.Parse</a></td></tr><tr><td class="src">parseOptions</td><td class="module"><a href="HoYo-CLI-Parse.html#v:parseOptions" title="HoYo.CLI.Parse">HoYo.CLI.Parse</a></td></tr><tr><td class="src">Print</td><td class="module"><a href="HoYo-Internal-Types.html#v:Print" title="HoYo.Internal.Types">HoYo.Internal.Types</a>, <a href="HoYo-Command.html#v:Print" title="HoYo.Command">HoYo.Command</a>, HoYo</td></tr><tr><td class="src">printStderr</td><td class="module"><a href="HoYo-Internal-Utils.html#v:printStderr" title="HoYo.Internal.Utils">HoYo.Internal.Utils</a>, <a href="HoYo.html#v:printStderr" title="HoYo">HoYo</a></td></tr><tr><td class="src">printStdout</td><td class="module"><a href="HoYo-Internal-Utils.html#v:printStdout" title="HoYo.Internal.Utils">HoYo.Internal.Utils</a>, <a href="HoYo.html#v:printStdout" title="HoYo">HoYo</a></td></tr></table></div></div><div id="footer"><p>Produced by <a href="http://www.haskell.org/haddock/">Haddock</a> version 2.24.2</p></div></body></html>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd"><html xmlns="http://www.w3.org/1999/xhtml"><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8" /><meta name="viewport" content="width=device-width, initial-scale=1" /><title>hoyo-0.3.0.0: Bookmark directories for cd (Index - P)</title><link href="linuwial.css" rel="stylesheet" type="text/css" title="Linuwial" /><link rel="stylesheet" type="text/css" href="quick-jump.css" /><link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=PT+Sans:400,400i,700" /><script src="haddock-bundle.min.js" async="async" type="text/javascript"></script><script type="text/x-mathjax-config">MathJax.Hub.Config({ tex2jax: { processClass: "mathjax", ignoreClass: ".*" } });</script><script src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-AMS-MML_HTMLorMML" type="text/javascript"></script></head><body><div id="package-header"><span class="caption">hoyo-0.3.0.0: Bookmark directories for cd</span><ul class="links" id="page-menu"><li><a href="index.html">Contents</a></li><li><a href="doc-index.html">Index</a></li></ul></div><div id="content"><div id="alphabet"><ul><li><a href="doc-index-A.html">A</a></li><li><a href="doc-index-B.html">B</a></li><li><a href="doc-index-C.html">C</a></li><li><a href="doc-index-D.html">D</a></li><li><a href="doc-index-E.html">E</a></li><li><a href="doc-index-F.html">F</a></li><li><a href="doc-index-G.html">G</a></li><li><a href="doc-index-H.html">H</a></li><li><a href="doc-index-I.html">I</a></li><li><a href="doc-index-L.html">L</a></li><li><a href="doc-index-M.html">M</a></li><li><a href="doc-index-N.html">N</a></li><li><a href="doc-index-O.html">O</a></li><li><a href="doc-index-P.html">P</a></li><li><a href="doc-index-R.html">R</a></li><li><a href="doc-index-S.html">S</a></li><li><a href="doc-index-T.html">T</a></li><li><a href="doc-index-U.html">U</a></li><li><a href="doc-index-V.html">V</a></li><li><a href="doc-index-W.html">W</a></li><li><a href="doc-index-95.html">_</a></li><li><a href="doc-index-All.html">All</a></li></ul></div><div id="index"><p class="caption">Index - P</p><table><tr><td class="src">parseCommand</td><td class="module"><a href="HoYo-Internal-Parse.html#v:parseCommand" title="HoYo.Internal.Parse">HoYo.Internal.Parse</a>, <a href="HoYo-CLI-Parse.html#v:parseCommand" title="HoYo.CLI.Parse">HoYo.CLI.Parse</a></td></tr><tr><td class="src">parseOptions</td><td class="module"><a href="HoYo-Internal-Parse.html#v:parseOptions" title="HoYo.Internal.Parse">HoYo.Internal.Parse</a>, <a href="HoYo-CLI-Parse.html#v:parseOptions" title="HoYo.CLI.Parse">HoYo.CLI.Parse</a></td></tr><tr><td class="src">parseOverride</td><td class="module"><a href="HoYo-Internal-Parse.html#v:parseOverride" title="HoYo.Internal.Parse">HoYo.Internal.Parse</a></td></tr><tr><td class="src">parseOverrideDisplayCreationTime</td><td class="module"><a href="HoYo-Internal-Parse.html#v:parseOverrideDisplayCreationTime" title="HoYo.Internal.Parse">HoYo.Internal.Parse</a></td></tr><tr><td class="src">parseOverrideEnableClearing</td><td class="module"><a href="HoYo-Internal-Parse.html#v:parseOverrideEnableClearing" title="HoYo.Internal.Parse">HoYo.Internal.Parse</a></td></tr><tr><td class="src">parseOverrideEnableReset</td><td class="module"><a href="HoYo-Internal-Parse.html#v:parseOverrideEnableReset" title="HoYo.Internal.Parse">HoYo.Internal.Parse</a></td></tr><tr><td class="src">parseOverrideFailOnError</td><td class="module"><a href="HoYo-Internal-Parse.html#v:parseOverrideFailOnError" title="HoYo.Internal.Parse">HoYo.Internal.Parse</a></td></tr><tr><td class="src">Print</td><td class="module"><a href="HoYo-Internal-Types.html#v:Print" title="HoYo.Internal.Types">HoYo.Internal.Types</a>, <a href="HoYo-Command.html#v:Print" title="HoYo.Command">HoYo.Command</a>, HoYo</td></tr><tr><td class="src">printStderr</td><td class="module"><a href="HoYo-Internal-Utils.html#v:printStderr" title="HoYo.Internal.Utils">HoYo.Internal.Utils</a>, <a href="HoYo.html#v:printStderr" title="HoYo">HoYo</a></td></tr><tr><td class="src">printStdout</td><td class="module"><a href="HoYo-Internal-Utils.html#v:printStdout" title="HoYo.Internal.Utils">HoYo.Internal.Utils</a>, <a href="HoYo.html#v:printStdout" title="HoYo">HoYo</a></td></tr></table></div></div><div id="footer"><p>Produced by <a href="http://www.haskell.org/haddock/">Haddock</a> version 2.24.2</p></div></body></html>
