+++ docs/index.html
@@ -1,3 +1,3 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd"><html xmlns="http://www.w3.org/1999/xhtml"><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8" /><meta name="viewport" content="width=device-width, initial-scale=1" /><title>hoyo-0.3.0.0: Bookmark directories for cd</title><link href="linuwial.css" rel="stylesheet" type="text/css" title="Linuwial" /><link rel="stylesheet" type="text/css" href="quick-jump.css" /><link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=PT+Sans:400,400i,700" /><script src="haddock-bundle.min.js" async="async" type="text/javascript"></script><script type="text/x-mathjax-config">MathJax.Hub.Config({ tex2jax: { processClass: "mathjax", ignoreClass: ".*" } });</script><script src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-AMS-MML_HTMLorMML" type="text/javascript"></script></head><body><div id="package-header"><span class="caption">hoyo-0.3.0.0: Bookmark directories for cd</span><ul class="links" id="page-menu"><li><a href="index.html">Contents</a></li><li><a href="doc-index.html">Index</a></li></ul></div><div id="content"><div id="description"><h1>hoyo-0.3.0.0: Bookmark directories for cd</h1><div class="doc"><p>hoyo is a command-line utility enabling easy bookmarking
 of directory for quick cd-like behaviour.</p><p>See the <a href="https://github.com/fpringle/hoyo#readme">github readme</a>
-for more information.</p></div></div><div id="module-list"><p class="caption">Modules</p><ul><li><span class="module"><span class="details-toggle-control details-toggle" data-details-id="n.1">&nbsp;</span><a href="HoYo.html">HoYo</a></span><details id="n.1" open="open"><summary class="hide-when-js-enabled">Submodules</summary><ul><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Bookmark.html">HoYo.Bookmark</a></span></li><li><span class="module details-toggle-control details-toggle" data-details-id="n.1.2">CLI</span><details id="n.1.2" open="open"><summary class="hide-when-js-enabled">Submodules</summary><ul><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-CLI-Complete.html">HoYo.CLI.Complete</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-CLI-Parse.html">HoYo.CLI.Parse</a></span></li></ul></details></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Command.html">HoYo.Command</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Config.html">HoYo.Config</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Env.html">HoYo.Env</a></span></li><li><span class="module details-toggle-control details-toggle" data-details-id="n.1.6">Internal</span><details id="n.1.6" open="open"><summary class="hide-when-js-enabled">Submodules</summary><ul><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Bookmark.html">HoYo.Internal.Bookmark</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Command.html">HoYo.Internal.Command</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Config.html">HoYo.Internal.Config</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Env.html">HoYo.Internal.Env</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Types.html">HoYo.Internal.Types</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Utils.html">HoYo.Internal.Utils</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Version.html">HoYo.Internal.Version</a></span></li></ul></details></li></ul></details></li></ul></div></div><div id="footer"><p>Produced by <a href="http://www.haskell.org/haddock/">Haddock</a> version 2.24.2</p></div></body></html>
+for more information.</p></div></div><div id="module-list"><p class="caption">Modules</p><ul><li><span class="module"><span class="details-toggle-control details-toggle" data-details-id="n.1">&nbsp;</span><a href="HoYo.html">HoYo</a></span><details id="n.1" open="open"><summary class="hide-when-js-enabled">Submodules</summary><ul><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Bookmark.html">HoYo.Bookmark</a></span></li><li><span class="module details-toggle-control details-toggle" data-details-id="n.1.2">CLI</span><details id="n.1.2" open="open"><summary class="hide-when-js-enabled">Submodules</summary><ul><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-CLI-Complete.html">HoYo.CLI.Complete</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-CLI-Parse.html">HoYo.CLI.Parse</a></span></li></ul></details></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Command.html">HoYo.Command</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Config.html">HoYo.Config</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Env.html">HoYo.Env</a></span></li><li><span class="module details-toggle-control details-toggle" data-details-id="n.1.6">Internal</span><details id="n.1.6" open="open"><summary class="hide-when-js-enabled">Submodules</summary><ul><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Bookmark.html">HoYo.Internal.Bookmark</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Command.html">HoYo.Internal.Command</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Config.html">HoYo.Internal.Config</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Env.html">HoYo.Internal.Env</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Parse.html">HoYo.Internal.Parse</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Types.html">HoYo.Internal.Types</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Utils.html">HoYo.Internal.Utils</a></span></li><li><span class="module"><span class="noexpander">&nbsp;</span><a href="HoYo-Internal-Version.html">HoYo.Internal.Version</a></span></li></ul></details></li></ul></details></li></ul></div></div><div id="footer"><p>Produced by <a href="http://www.haskell.org/haddock/">Haddock</a> version 2.24.2</p></div></body></html>
