+++ docs/src/HoYo.Command.html
@@ -26,7 +26,7 @@ command-line program.
 </span><span id="line-26"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#runDefaultCommand"><span class="hs-identifier">runDefaultCommand</span></a></span><span>
 </span><span id="line-27"></span><span>
 </span><span id="line-28"></span><span>  </span><span class="annot"><span class="hs-comment">-- * Types</span></span><span>
-</span><span id="line-29"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#Options"><span class="hs-identifier">Options</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
+</span><span id="line-29"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#Options"><span class="hs-identifier">Options</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
 </span><span id="line-30"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#Command"><span class="hs-identifier">Command</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
 </span><span id="line-31"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#AddOptions"><span class="hs-identifier">AddOptions</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
 </span><span id="line-32"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#MoveOptions"><span class="hs-identifier">MoveOptions</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
@@ -40,15 +40,15 @@ command-line program.
 </span><span id="line-40"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#ConfigAddDefaultOptions"><span class="hs-identifier">ConfigAddDefaultOptions</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
 </span><span id="line-41"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#ConfigCommand"><span class="hs-identifier">ConfigCommand</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
 </span><span id="line-42"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#CheckOptions"><span class="hs-identifier">CheckOptions</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
-</span><span id="line-43"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#GlobalOptions"><span class="hs-identifier">GlobalOptions</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
+</span><span id="line-43"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#GlobalOptions"><span class="hs-identifier">GlobalOptions</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
 </span><span id="line-44"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#defaultGlobalOptions"><span class="hs-identifier">defaultGlobalOptions</span></a></span><span>
-</span><span id="line-45"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#OverrideOptions"><span class="hs-identifier">OverrideOptions</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
+</span><span id="line-45"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#OverrideOptions"><span class="hs-identifier">OverrideOptions</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
 </span><span id="line-46"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#defaultOverrideOptions"><span class="hs-identifier">defaultOverrideOptions</span></a></span><span>
 </span><span id="line-47"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#overrideConfig"><span class="hs-identifier">overrideConfig</span></a></span><span>
 </span><span id="line-48"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#overrideEnv"><span class="hs-identifier">overrideEnv</span></a></span><span>
 </span><span id="line-49"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#verifyOverrides"><span class="hs-identifier">verifyOverrides</span></a></span><span>
 </span><span id="line-50"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#combOverride"><span class="hs-identifier">combOverride</span></a></span><span>
-</span><span id="line-51"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html#MaybeOverride"><span class="hs-identifier">MaybeOverride</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
+</span><span id="line-51"></span><span>  </span><span class="hs-special">,</span><span> </span><span class="annot"><a href="HoYo.Internal.Types.html#MaybeOverride"><span class="hs-identifier">MaybeOverride</span></a></span><span> </span><span class="hs-special">(</span><span class="hs-glyph">..</span><span class="hs-special">)</span><span>
 </span><span id="line-52"></span><span>  </span><span class="hs-special">)</span><span> </span><span class="hs-keyword">where</span><span>
 </span><span id="line-53"></span><span>
 </span><span id="line-54"></span><span class="hs-keyword">import</span><span> </span><span class="annot"><a href="HoYo.Internal.Command.html"><span class="hs-identifier">HoYo.Internal.Command</span></a></span><span>
