+++ test/HoYo/Test/Gen.hs
@@ -65,9 +65,6 @@ instance Arbitrary BookmarkSearchTerm where
     , SearchName <$> textNotNumber
     ]
 
-genCommandText :: Gen T.Text
-genCommandText = pure ""
-
 instance Arbitrary (ConfigValue 'TBool) where
   arbitrary = BoolV <$> arbitrary
 
@@ -75,7 +72,7 @@ instance Arbitrary (ConfigValue 'TDefaultBookmark) where
   arbitrary = DefaultBookmarkV <$> arbitrary
 
 instance Arbitrary (ConfigValue 'TCommand) where
-  arbitrary = CommandV <$> genCommandText
+  arbitrary = CommandV <$> arbitrary
 
 instance Arbitrary (ConfigValue t) => Arbitrary (ConfigValue ('TMaybe t)) where
   arbitrary = MaybeV <$> arbitrary
